+++ packages/antd-management-fast-component/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Change Log
 
+## 2.6.139
+
+### Patch Changes
+
+- update
+
 ## 2.6.138
 
 ### Patch Changes