+++ packages/antd-management-fast-design-playground/CHANGELOG.md
@@ -1,5 +1,19 @@
 # antd-management-fast-design-playground
 
+## 1.0.101
+
+### Patch Changes
+
+- update
+- Updated dependencies
+  - antd-management-fast-component@2.6.139
+
+## 1.0.100
+
+### Patch Changes
+
+- update
+
 ## 1.0.99
 
 ### Patch Changes